requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord.py==1.5.0
feedparser==6.0.1
fenixedu==1.0.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
psycopg2==2.8.6
python-dotenv==0.14.0
requests==2.24.0
sgmllib3k==1.0.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
